Stop causing wl-paste `No selection` error
I discovered that writing to the `+` register caused wl-paste to give a
`No selection` error after quitting Vim. That error goes away if Vim is
run without connecting to the X server (i.e. `vim -X`), but then my
paste remap no longer works. (Presumably because it essentially maps to
`"+p`, which tries to use the X clipboard integration but can't because
it's not connected to the X server.) Running a build of Vim without X
(and therefore clipboard) support doesn't work either because then the
`+` register isn't recognized.

As a result, I switched to using the unnamed register (`"`) as an
intermediary in the paste mappings. This gets rid of the `No selection`
error and allows the plugin to work whether Vim is run as normal or
without connecting to X.