Remove dependency on `+clipboard` feature
On Vim builds without `+clipboard`, the `+` register doesn't work for
yanking. My solution is to map `"+` to `"w` and send the `w` register to
the Wayland clipboard as well. (This only occurs when the `+clipboard`
feature is missing.) This feature can be disabled by setting the
`g:wayland_clipboard_no_plus_to_w` variable.