Enable w reg clobbering if xwayland isn't available (no $DISPLAY)
The + register in vim isn't available if it isn't built with the
`clipboard` feature. We work around this by detecting it, and
remapping the + register to w.

But the + register can also be unavailable if vim is unable to connect
to an X server.

Try to detect the lack of X server by checking the $DISPLAY environment
variable (which X applications use to connect to the right X server, but
isn't used by wayland), and also use the w register in that case.